==> 01_Data_Acquisition_and_Cleaning.ipynb <==
# Creating a Dataset for American Movies (1970-2023) from Scratch

In this notebook, we undertake the task of creating a comprehensive dataset for American movies released between 1970 and 2024. The process involves several key stages:

## 1. Data Collection

### 1.1 Retrieving Movie Titles, Cast and Crew, and Release Year from American Movies

In this stage, we dynamically construct Wikipedia URLs for each year from 1970 to 2024 to gather essential information such as movie titles, cast and crew details, and release years.

## 2. Data Ingestion

### 2.1 Loading Data into the Notebook Environment

Once the necessary data is collected, we proceed to ingest it into the notebook environment. This step involves reading the data The Movie Data Base (TMDb) API.

## 3. Data Processing

### 3.1 Cleaning and Preprocessing the Data

With the data loaded into the notebook, we perform essential preprocessing tasks such as duplicate removal, missing value imputation, and text cleaning. These steps ensure the data's quality and readiness for further analysis.

## 4. Data Cleaning

### 4.1 Ensuring Data Quality and Consistency

In this final stage, we conduct quality assurance checks to verify the integrity and consistency of the dataset. This involves validating data ranges, performing consistency checks, and implementing error handling mechanisms.



import os
import requests
import numpy as np
import pandas as pd
import time
from tqdm import tqdm
from dotenv import load_dotenv

# We construct dynamically Wikipedia URLs for each year from 1970 to 2023.
def get_films(year):

    url = f"https://en.wikipedia.org/wiki/List_of_American_films_of_{year}"
    df = pd.read_html(url, header=0)

    # Take relevant tables only
    df = pd.concat(df[2:6])
    df.columns = df.columns.str.lower()
    if 'ref.' in df.columns:
        df = df.drop(['opening', 'opening.1', 'production company', 'ref.'], axis=1)
    else:
        df = df.drop(['opening', 'opening.1', 'production company'], axis=1)
    df['year'] = year
    return df

df = pd.DataFrame()
for year in range(1970,2017):
    print(year)
    df = pd.concat([df, get_films(year)])

df.head()

df = df.drop(labels=['unnamed: 5', '$217,581,231'], axis=1)
df

For 2017 there is a typo in wikitable class, so let's processing separately.

from bs4 import BeautifulSoup
from io import StringIO
import requests

url = "https://en.wikipedia.org/wiki/List_of_American_films_of_2017"
response = requests.get(url)
soup = BeautifulSoup(response.text, 'html.parser')

# Find the problematic cell and set the rowspan attribute to '1'
problematic_cell = soup.find('td', {'style': 'text-align:center; background:#dfc7f7', 'rowspan': 'q'})
if problematic_cell:
    problematic_cell['rowspan'] = '1'

# Get the updated HTML content
updated_html = str(soup)

# Read the HTML table into a DataFrame
df_list = pd.read_html(StringIO(updated_html))
df_list = pd.concat(df_list[2:6])
df_list.columns = df_list.columns.str.lower()
df_list = df_list.drop(['opening', 'opening.1', 'production company', 'ref.'], axis=1)
df_list['year'] = 2017
df_list

df = pd.concat([df, df_list])
df

Continue up to 2023

for year in range(2018,2024):
    print(year)
    df = pd.concat([df, get_films(year)])

df

Before saving to a file, let's be sure that there are no null values in the title and year fields.

df.isna().sum()

df.dropna(subset=['title'], inplace=True)

len(df)

df.reset_index(drop=True, inplace=True)
df

df.to_csv('./data/movies_1970-2023.csv')

## 2. Data Ingestion. Retrieving movie IDs from "The Movie Data Base"

In the second phase of this project, we shifted our focus to Data Ingestion, specifically aimed at retrieving unique movie identifiers (IDs) from "The Movie Database" (TMDb). This step was crucial for augmenting our dataset with additional information and establishing a comprehensive linkage between movies. By interfacing with TMDb's API, we accessed movie IDs, which serve as universal identifiers across various platforms. This integration not only enriched our dataset but also opened avenues for obtaining detailed metadata and ratings, setting the stage for a more robust and insightful analysis of movies.

import numpy as np

def parsing_fun(row, data_json):
    # Initialize an empty list to store data for matching movies
    matched_movies_data = []

    # Check if 'total_results' list is not empty
    if data_json.get('total_results'):
        # Iterate through the 'results' list in the JSON data
        for movie in data_json.get('results'):
            # Extract the release date
            release_date = movie.get('release_date', '')

            # Check if there is a release data
            if release_date == '':
                release_date = '0'

            # Check if the movie is from the specified year and has a title similar to the provided title
            delta = np.abs(int(release_date.split('-')[0]) - row.get('year'))
            if release_date and (delta <= 1):
                matched_movies_data.append({
                    "id": int(movie.get('id', np.nan)),
                    "genre_ids": movie.get('genre_ids', np.nan),
                    "original_title": movie.get('original_title', ''),
                    "title": row.get('title', ''),
                    "cast and crew": row.get('cast and crew', ''),
                    "year": row.get('year', np.nan),
                    "poster_path": movie.get('poster_path', ''),
                })

    else:
        # Append NaN values if 'results' list is empty
        matched_movies_data.append({
            "id": np.nan,
            "genre_ids": np.nan,
            "original_title": '',
            "title": row.get('title', ''),
            "cast and crew": row.get('cast and crew', ''),
            "year": row.get('year', np.nan),
            "poster_path": '',
        })

    # Convert the list of dictionaries to a DataFrame
    matched_movies_df = pd.DataFrame(matched_movies_data)

    return matched_movies_df

load_dotenv()

my_api_key = os.environ.get('tmdb_api_key')

# Create an empty DataFrame with the desired columns
header_df = pd.DataFrame(columns=['id', 'genre_ids', 'original_title', 'title', 'cast and crew', 'year', 'poster_path'])

FILE_NAME = './data/movie_ids_.csv'
# Save the header DataFrame to a CSV file
header_df.to_csv(FILE_NAME, index=False)

for _, row in tqdm(df.iterrows(), total=len(df), desc="Processing"):

    title = row['title']
    year = row['year']
    
    while True:
        try:
            response = requests.get(f"https://api.themoviedb.org/3/search/movie?api_key={my_api_key}&query={title}")
            response.raise_for_status()  # Raise HTTPError for bad responses (4xx and 5xx)
            
            if response.status_code == 200:
                data_json = response.json()

                current_data = parsing_fun(row, data_json)
                
                # Append to the list
                current_data.to_csv(FILE_NAME, mode='a', header=False, index=False)
                
                break  # Exit the loop if the request is successful
            
        except requests.exceptions.RequestException as e:
            print(f"Error for {title}: {e}")
            # Handle specific exceptions if needed
        
        if response.status_code == 429:
            print(f"Rate limited. Waiting for 10 seconds before retrying...")
            time.sleep(10)
        else:
            break  # Exit the loop if it's not a rate-limiting issue

    time.sleep(0.3)  # Optional: Add a small delay to avoid rate limiting

## 3. Data Processing: Step 1 - Deleting Empty Fields

Given that the movie was not found in the TMDb database, let's proceed by deleting any row where the first three fields are empty. Most of these entries originate from countries other than the USA.

```
,,,Carry On Up the Jungle (United Kingdom),"Gerald Thomas"
```

df = pd.read_csv('./data/movies_id.csv')
df

df.info()

df.dropna(subset='id', inplace=True)
df.info()

Let's manually retrieve the cast and crew information for three movies from their respective Wikipedia pages. Below are the details obtained for each movie:

df[df.cast_and_crew.isna()]

df.loc[205, 'cast_and_crew'] = "Silvio Narizzano (director); Ray Galton and Alan Simpson (screenplay); 	Richard Attenborough, Lee Remick, Hywel Bennett, Milo O'Shea, Roy Holder"

df.loc[2306, 'cast_and_crew'] = "Donald W. Thompson (director); Donald W. Thompson, Russell S. Doughten Jr. (story/screenplay); William Wellman Jr., Susan Plumb, Patty Dunning"

df.loc[10526, 'cast_and_crew'] = "Adam Silver, Ben Hayflick (director); Patrick Muldoon, Bo Svenson"

## 4. Cleaning Processing. Step 2:

To ensure data consistency, we will check for any duplicates in the `original_title` and `title fields`. In case of duplicates, we will manually verify which one is correct based on their respective IDs.

    `28681,"[12, 27]",Equinox,Equinox,"Jack Woods (director/screenplay);` (correct)

    `389035,"[18, 878]",Equinozio,Equinox,"Jack Woods (director/screenplay);` (incorrect)

- 

# Create 'status' column to False
df_director = df[df.duplicated(subset='cast_and_crew', keep=False)]
df_director['status'] = False
df_director

Since we have duplicates in its `original title` field, let's find the director from API and compare to the `cast_and_crew` colum.

from fuzzywuzzy import fuzz

def is_duplicate(index, row, data_json):
    matched_id = []
    crew = data_json.get('crew')

    for person in crew:
        job = person.get('job', '')
        if job == 'Director':
            name = person.get('name')
            similarity_score = fuzz.partial_ratio(name.lower(), row.cast_and_crew.lower())

            status = similarity_score >= 80

            matched_id.append({
                'index': index,
                'id': int(row.get('id')),
                'original_title': row.get('original_title'),
                'name': name,
                'crew': row.cast_and_crew.split(';')[0],
                'status': status,
            })

    return pd.DataFrame(matched_id)

load_dotenv()
my_api_key = os.environ.get('tmdb_api_key')

header_df = pd.DataFrame(columns=['index', 'id', 'original_title', 'name', 'crew', 'status'])
FILE_NAME = 'duplicates_.csv'
header_df.to_csv(FILE_NAME, index=False)

i=0
for index, row in tqdm(df_duplicated.iterrows(), total=len(df_duplicated), desc="Processing"):

    movie_id = int(row['id'])
    title = row.title

    while True:
        try:
            response = requests.get(f"https://api.themoviedb.org/3/movie/{movie_id}/credits?api_key={my_api_key}")
            response.raise_for_status()

            if response.status_code == 200:
                data_json = response.json()

                result = is_duplicate(index, row, data_json)
                result.to_csv(FILE_NAME, mode='a', header=False, index=False)
                break
            
        except requests.exceptions.RequestException as e:
            print(f"Error for {title}: {e}")
        
        if response.status_code == 429:
            print("Waiting for 10 seconds before retrying...")
            time.sleep(10)
        else:
            break

    time.sleep(0.3)


df_duplicate = pd.read_csv('./data/duplicates.csv', index_col=0)
df_duplicate

# Identify rows to keep
keep_by_index = df_duplicate.loc[df_duplicate['status'] == True].index

# Update 'status'
df_director.loc[df_director.index.isin(keep_by_index), 'status'] = True
df_director

df_director.loc[df_director['status'] == False]

df[df.duplicated(subset=['title', 'year'],  keep=False)]

keep_by_index = df_director.loc[df_director['status'] == False].index

# Delete repeated rows
df_cleaned = df[~df.index.isin(keep_by_index)]

df_cleaned = df_cleaned.drop_duplicates(subset=['title', 'year'], keep='first')

df_cleaned = df_cleaned.drop_duplicates(subset='id')

# Ids are integers
df_cleaned['id'] = df_cleaned['id'].astype(int)

df_cleaned.to_csv('movies_cleaned_.csv', index=False)

df_cleaned.isna().sum()

## 5. Retrieve ratings

In the fifth phase of our data-driven project, we delved into the process of retrieving movie ratings. This step involved accessing and incorporating user-generated ratings.

Leveraging the ratings dataset, we aimed to gauge the popularity, viewer preferences, and overall acclaim for each movie within our curated collection. This wealth of rating information not only enhances the depth of our dataset but also forms the foundation for building a recommender system that tailors movie suggestions based on user preferences.

import numpy as np

def get_ratings(row, data_json):
    matched_id = []

    if data_json.get('total_results'):
        for author in data_json.get('results'):
            id = author.get('id')
            author_details = author.get('author_details')
            rating = author_details.get('rating', np.nan)

            matched_id.append({
                "movieId": int(row.get('id', 0)),
                "userId": id,
                "rating": rating,
            })

    else:
        matched_id.append({
                "movieId": int(row.get('id', 0)),
                "userId": '',
                "rating": '',
            })
        
    matched_id_df = pd.DataFrame(matched_id)

    return matched_id_df

import requests
from tqdm import tqdm
import time

header_df = pd.DataFrame(columns=['movieId', 'userId', 'rating'])

FILE_NAME = './data/ratings_.csv'
header_df.to_csv(FILE_NAME, index=False)

for _, row in tqdm(df_cleaned.iterrows(), total=len(df_cleaned), desc='Processing'):
    
    movieId = int(row.get('id', 0))

    while True:
        try:
            response = requests.get(f"https://api.themoviedb.org/3/movie/{movieId}/reviews?api_key={my_api_key}")
            response.raise_for_status()

            if response.status_code == 200:
                data_json = response.json()

                current_data = get_ratings(row, data_json)

                current_data.to_csv(FILE_NAME, mode='a', header=False, index=False)

                break

        except requests.exceptions.RequestException as e:
            print(f"Error for {int(movieId)}: {e}")
        
        if response.status_code == 429:
            print(f"Rate limited. Waiting for 10 seconds before retrying...")
            time.sleep(10)
        else:
            break
    
    time.sleep(0.3)

df_rating = pd.read_csv('./data/ratings.csv')
df_rating

df_rating.isna().sum()

df_rating.dropna(inplace=True)

len(rating_cleaned)

rating_cleaned = rating_cleaned.drop_duplicates()

# Keep index from movies_cleaned.csv
rating_cleaned = rating_cleaned[rating_cleaned['movieId'].isin(df_cleaned['id'])]

## 6. Anonymizing Users

In order to protect user privacy and comply with data protection regulations, we will anonymize user data.

# Create a mapping between original user IDs and anonymous IDs
unique_users = rating_cleaned['userId'].unique()

user_id_mapping = {user_id: f"{idx}" for idx, user_id in enumerate(unique_users)}

rating_cleaned['userId'] = rating_cleaned['userId'].map(user_id_mapping)

rating_cleaned.to_csv('ratings_cleaned.csv', index=False)

## 7. Genre Names

In this section, we create a mapping from genre IDs to genre names.

df.genre_ids

import json

with open('genres.json', 'r') as fp:
    genres_data = json.load(fp)

genres_mapping = {genre.get('id'): genre.get('name') for genre in genres_data.get('genres')}

from ast import literal_eval

df['genre_ids'] = df['genre_ids'].apply(literal_eval)

df['genre'] = df['genre_ids'].apply(lambda ids: ', '.join(genres_mapping[genre_id] for genre_id in ids))

df.drop('genre_ids', axis=1, inplace=True)

df.drop('original_title', axis=1, inplace=True)

df[df['genre']==''] ='unknown'

df.to_csv('movies.csv', index=False)
